permission and graphs
given pemission to controllers and pages and build the graphs

+++ MovieMonster/Controllers/SalesController.cs
@@ -8,6 +8,7 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using MovieMonster.Data;
 using MovieMonster.Models;
+using Accord.MachineLearning.Rules;
 
 namespace MovieMonster.Controllers
 {
@@ -20,15 +21,15 @@ namespace MovieMonster.Controllers
         {
             _context = context;
         }
-
         // GET: Sales
+        [Authorize(Roles = "Admin")]
         public async Task<IActionResult> Index()
         {
             // include Movies from MovieSale(Movies)
             var movieMonsteContext = _context.Sale.Include(s => s.Customer).Include(ms => ms.Movies).Include("Movies.Movie");
             return View(await movieMonsteContext.ToListAsync());
         }
-
+        [Authorize(Roles = "User")]
         // GET: Sales/Details/5
         public async Task<IActionResult> Details(string id)
         {
@@ -43,6 +44,7 @@ namespace MovieMonster.Controllers
             {
                 return NotFound();
             }
+
 
             return View(sale);
         }
@@ -59,13 +61,14 @@ namespace MovieMonster.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         //[HttpPost]
         //[ValidateAntiForgeryToken]
+        [Authorize(Roles = "User")]
         public async Task<IActionResult> Create(string MovieID, [Bind("SaleID,CustomerID,Purchased,TotalPrice")] Sale sale)
         {
             if (ModelState.IsValid)
             {
                 _context.Add(sale);
                 await _context.SaveChangesAsync();
-                return RedirectToAction("Create", "MovieSale", new
+                return RedirectToAction("Create", "MovieSales", new
                 {
                     SaleID = sale.SaleID,
                     MovieID = MovieID
@@ -76,6 +79,7 @@ namespace MovieMonster.Controllers
         }
 
         // GET: Sales/Edit/5
+        [Authorize(Roles = "User")]
         public async Task<IActionResult> Edit(string id)
         {
             if (id == null)
@@ -95,6 +99,7 @@ namespace MovieMonster.Controllers
         // POST: Sales/Edit/5
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
+        [Authorize(Roles = "User")]
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(string id, [Bind("SaleID,CustomerID,Purchased,TotalPrice")] Sale sale)
@@ -147,14 +152,19 @@ namespace MovieMonster.Controllers
         }
 
         // POST: Sales/Delete/5
+        [Authorize(Roles = "User")]
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(string id)
         {
             var sale = await _context.Sale.FindAsync(id);
-            _context.Sale.Remove(sale);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            if (sale.Purchased == false)
+            {
+                _context.Sale.Remove(sale);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            return RedirectToAction("Index", "Home");
         }
 
         private bool SaleExists(string id)
@@ -199,7 +209,7 @@ namespace MovieMonster.Controllers
             }
             return View("Index", cart);
         }
-
+        [Authorize(Roles = "User")]
         public async Task<IActionResult> Purchase(string id)
         {
             if (id == null)
@@ -231,11 +241,13 @@ namespace MovieMonster.Controllers
             }
             return RedirectToAction("Index", "Home");
         }
-        //Direction to the graph view =  /Sales/PieChart
-        public IActionResult PieChart()
+
+        //Direction to the graph view =  /Sales/Statistics
+        public ActionResult Statistics()
         {
-            return View("~/Views/Graphs/PieChart.cshtml");
+            return View();
         }
+
         // Join And groupBy query result, sent as array to the view by Ajax
         public JsonResult QuantityJson()
         {
@@ -251,13 +263,69 @@ namespace MovieMonster.Controllers
                                           Quantity = finalTable.Sum(q => q.Quantity)
                                       }
                                       );
+            var resultTableAsList = titleQuantityTable.OrderByDescending(x => x.Quantity).ToList();
+            return Json(resultTableAsList);
 
-            var resultTableAsArray = titleQuantityTable.ToList();
-
-            //var json = JsonConvert.SerializeObject(resultTableAsArray);
-
-            return Json(resultTableAsArray);
-            
         }
+
+        public void Predict(string SaleID, string CustomerID)
+        {
+            Dictionary<string, int> mapActors = new Dictionary<string, int>();
+            SortedSet<int>[] dataset = SalesTransactionConverter(mapActors);
+            Apriori apriori = new Apriori(threshold: 3, confidence: 0);
+            AssociationRuleMatcher<int> classifier = apriori.Learn(dataset);
+            var sale = _context.Sale.Include(s => s.Movies).Include("Movies.Movie").FirstOrDefault(s => s.SaleID == SaleID);
+
+            HashSet<string> actorsSet = new HashSet<string>();
+            List<int> sample = new List<int>();
+            foreach (var movie in sale.Movies)
+            {
+                string[] actors = movie.Movie.Actors.Split(",");
+                foreach (var actor in actors)
+                {
+                    if (!actorsSet.Contains(actor))
+                    {
+                        actorsSet.Add(actor);
+                        sample.Add(mapActors.GetValueOrDefault(actor));
+                    }
+                }
+            }
+            int[][] matches = classifier.Decide(sample.ToArray());
+        }
+
+        public SortedSet<int>[] SalesTransactionConverter(Dictionary<string, int> mapActors)
+        {
+            int counter = 0;
+            var sales = _context.Sale.Include(s => s.Movies).Include("Movies.Movie").ToList();
+
+            List<SortedSet<int>> datasetList = new List<SortedSet<int>>();
+            foreach (var sale in sales)
+            {
+                HashSet<string> actorsSet = new HashSet<string>();
+                SortedSet<int> actors = new SortedSet<int>();
+                foreach (var movie in sale.Movies)
+                {
+                    string[] temp = movie.Movie.Actors.Split(",");
+                    foreach (var actor in temp)
+                    {
+                        if (!actorsSet.Contains(actor))
+                            if (mapActors.ContainsKey(actor))
+                            {
+                                actors.Add(mapActors.GetValueOrDefault(actor));
+                            }
+                            else
+                            {
+                                counter++;
+                                mapActors.Add(actor, counter);
+                                actors.Add(counter);
+                            }
+                    }
+                }
+                datasetList.Add(actors);
+            }
+            SortedSet<int>[] dataset = datasetList.ToArray();
+            return dataset;
+        }
+
     }
 }
